# code/sonara_back_end/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sonara API",
    "description": "\nAPI do projeto **Sonara** — plataforma de conexão entre artistas e organizadores de eventos.\n\n## Autenticação\nAlgumas rotas exigem **Bearer Token** (JWT). Faça login em `POST /v1/sonara/usuario/login` e use o token retornado no header:\n`Authorization: Bearer <token>`\n\n## Fluxo principal\n1. Cadastre um usuário (artista ou organizador)\n2. Faça login para obter o token\n3. Crie eventos (organizador) ou candidate-se a eventos (artista)\n4. Gerencie candidaturas, contra-propostas e convites\n        ",
    "version": "1.0.10",
    "contact": {
      "name": "Davi de Almeida Santos"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Insira: Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sonara/usuario/organizador/{id}": {
      "get": {
        "tags": [
          "Usuário"
        ],
        "summary": "Buscar organizador pelo ID do usuário",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Organizador encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Organizador não encontrado"
          },
          "415": {
            "description": "Unsupported Media Type"
          }
        }
      }
    },
...
    "/v1/sonara/organizador/": {
      "get": {
        "tags": [
          "Organizador"
        ],
        "summary": "Listar todos os organizadores",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum organizador encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Organizador"
        ],
        "summary": "Cadastrar organizador",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Organizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Organizador criado"
          },
          "204": {
            "description": "No Content"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/organizador/{id}": {
      "get": {
        "tags": [
          "Organizador"
        ],
        "summary": "Buscar organizador por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do organizador"
          }
        ],
        "responses": {
          "200": {
            "description": "Organizador encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Organizador não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Organizador"
        ],
        "summary": "Atualizar organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do organizador"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Organizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organizador atualizado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Organizador não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Organizador"
        ],
        "summary": "Excluir organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do organizador"
          }
        ],
        "responses": {
          "200": {
            "description": "Organizador excluído"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Organizador não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/evento/": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Listar todos os eventos",
        "description": "Retorna todos os eventos cadastrados com fotos, endereço e organizador.",
        "responses": {
          "200": {
            "description": "Lista de eventos retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum evento encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Evento"
        ],
        "summary": "Criar evento",
        "description": "Cria evento com endereço e vínculo ao organizador.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos obrigatórios ausentes ou inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/evento/{id}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Buscar evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento"
        ],
        "summary": "Atualizar evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento"
        ],
        "summary": "Excluir evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento excluído com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/eventoOrganizador/": {
      "get": {
        "tags": [
          "Evento x Organizador"
        ],
        "summary": "Listar todos os vínculos evento-organizador",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Evento x Organizador"
        ],
        "summary": "Criar vínculo evento-organizador",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoOrganizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Vínculo criado"
          },
          "204": {
            "description": "No Content"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/eventoOrganizador/{id}": {
      "get": {
        "tags": [
          "Evento x Organizador"
        ],
        "summary": "Buscar vínculo evento-organizador por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Vínculo encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento x Organizador"
        ],
        "summary": "Atualizar vínculo evento-organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoOrganizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Atualizado"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento x Organizador"
        ],
        "summary": "Excluir vínculo evento-organizador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Excluído"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EventoCadastro": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Festival Sonara 2026"
        },
        "descricao": {
          "type": "string",
          "example": "Festival de música ao ar livre"
        },
        "local": {
          "type": "string",
          "example": "Parque Ibirapuera"
        },
        "data": {
          "type": "string",
          "example": "2026-08-15"
        },
        "hora_inicio": {
          "type": "string",
          "example": "18:00"
        },
        "hora_fim": {
          "type": "string",
          "example": "23:00"
        },
        "organizador_id": {
          "type": "number",
          "example": 1
        },
        "cep": {
          "type": "string",
          "example": "04094-050"
        },
        "cidade": {
          "type": "string",
          "example": "São Paulo"
        },
        "estado": {
          "type": "string",
          "example": "SP"
        },
        "logradouro": {
          "type": "string",
          "example": "Av. Pedro Álvares Cabral"
        },
        "numero": {
          "type": "string",
          "example": "S/N"
        },
        "bairro": {
          "type": "string",
          "example": "Vila Mariana"
        },
        "complemento": {
          "type": "string",
          "example": "Portão 3"
        }
      },
      "required": [
        "nome",
        "descricao",
        "local",
        "data",
        "hora_inicio",
        "hora_fim",
        "organizador_id",
        "cep",
        "cidade",
        "estado",
        "logradouro",
        "numero",
        "bairro"
      ]
    },
...
    "Organizador": {
      "type": "object",
      "properties": {
        "usuario_id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "usuario_id"
      ]
    },
...
    "EventoOrganizador": {
      "type": "object",
      "properties": {
        "evento_id": {
          "type": "number",
          "example": 1
        },
        "organizador_id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "evento_id",
        "organizador_id"
      ]
    },
...
  }
}